third_party/WebKit/LayoutTests/imported/web-platform-tests/html/webappapis/animation-frames/idlharness-expected.txt

CONSOLE MESSAGE: line 198: callback not yet supported
This is a testharness.js-based test.
FAIL Window interface: operation requestAnimationFrame(FrameRequestCallback) assert_own_property: global object missing non-static operation expected property "requestAnimationFrame" missing
FAIL Window interface: operation cancelAnimationFrame(long) assert_own_property: global object missing non-static operation expected property "cancelAnimationFrame" missing
PASS Window must be primary interface of window 
FAIL Stringification of window assert_equals: class string of window expected "[object Window]" but got "[object global]"
FAIL Window interface: window must inherit property "requestAnimationFrame" with the proper type (0) assert_own_property: expected property "requestAnimationFrame" missing
FAIL Window interface: calling requestAnimationFrame(FrameRequestCallback) on window with too few arguments must throw TypeError assert_own_property: expected property "requestAnimationFrame" missing
FAIL Window interface: window must inherit property "cancelAnimationFrame" with the proper type (1) assert_own_property: expected property "cancelAnimationFrame" missing
FAIL Window interface: calling cancelAnimationFrame(long) on window with too few arguments must throw TypeError assert_own_property: expected property "cancelAnimationFrame" missing
Harness: the test ran to completion.
